RAVE 0.440
- Implementat DMenuEx amb una barra de desplaçament
- Solucionat bug de la captura, era pel WM_NCACTIVATE que retornava FALSE, ara retorna TRUE.
- Afegit menu Video -> Brillo amb una barra

+++ Reproductor de Audio y Video Extendido/VentanaOpcionesRAVE.cpp
@@ -112,7 +112,7 @@ void VentanaOpcionesRAVE::Crear(void) {
 		L"Reproducir siguiente",
 		L"Reporducir anterior"
 	};
-	for (int i = 0; i < App.TeclasRapidas.size(); i++) {
+	for (int i = 0; i < static_cast<int>(App.TeclasRapidas.size()); i++) {
 		EtiquetasTeclas[i].CrearEtiquetaEx(&MarcoGeneral, Textos[i], 10, 80 + (i * 25), 160, 20, ID_ETIQUETAS_TECLADO + i);
 		EtiquetasTeclas[i].Fuente = Fuente18Normal;
 		TeclasRapidas[i].Crear(&MarcoGeneral, 180, 80 + (i * 25), RC.right - 210, 20, ID_ASINGAR_TECLA, &App.TeclasRapidas[i]);
